@@ -52,13 +52,13 @@
   ],
   "dependencies": {
     "@loopback/authentication": "^10.1.3",
-    "@loopback/boot": "^6.1.2",
-    "@loopback/core": "^5.1.2",
-    "@loopback/repository": "^6.1.2",
-    "@loopback/rest": "^13.1.2",
-    "@loopback/rest-explorer": "^6.1.2",
+    "@loopback/boot": "^6.1.3",
+    "@loopback/core": "^5.1.3",
+    "@loopback/repository": "^6.1.3",
+    "@loopback/rest": "^13.1.3",
+    "@loopback/rest-explorer": "^6.1.3",
     "@loopback/security": "^0.10.3",
-    "@loopback/service-proxy": "^6.1.2",
+    "@loopback/service-proxy": "^6.1.3",
     "@sendgrid/mail": "^7.7.0",
     "bcrypt": "^5.1.1",
     "cloudinary": "^1.41.0",
@@ -75,19 +75,19 @@
     "twilio": "^4.17.0"
   },
   "devDependencies": {
-    "@loopback/build": "^10.1.2",
-    "@loopback/eslint-config": "^14.0.3",
-    "@loopback/testlab": "^6.1.2",
+    "@loopback/build": "^10.1.3",
+    "@loopback/eslint-config": "^14.0.4",
+    "@loopback/testlab": "^6.1.3",
     "@types/bcrypt": "^5.0.0",
     "@types/jsonwebtoken": "^9.0.2",
     "@types/multer": "^1.4.8",
-    "@types/node": "^16.18.40",
+    "@types/node": "^16.18.50",
     "@types/node-dijkstra": "^2.5.4",
     "@types/speakeasy": "^2.0.7",
     "@types/streamifier": "^0.1.0",
-    "eslint": "^8.47.0",
+    "eslint": "^8.48.0",
     "source-map-support": "^0.5.21",
     "tsc-watch": "^6.0.4",
-    "typescript": "~5.1.6"
+    "typescript": "~5.2.2"
   }
 }
