@@ -8,14 +8,10 @@ export * from './point.controller';
 export * from './pqrs.controller';
 export * from './role-user.controller';
 export * from './role.controller';
-export * from './route-point.controller';
-export * from './route-trip.controller';
-export * from './route.controller';
 export * from './trip-comment-trip.controller';
 export * from './trip-comment-user.controller';
 export * from './trip-comment.controller';
 export * from './trip-rating.controller';
-export * from './trip-route.controller';
 export * from './trip-trip-comment.controller';
 export * from './trip-user.controller';
 export * from './trip.controller';
@@ -24,6 +20,4 @@ export * from './user-trip-comment.controller';
 export * from './user.controller';
 export * from './vehicle-user.controller';
 export * from './vehicle.controller';
-export * from './route-point-route.controller';
-export * from './route-point-point.controller';
-export * from './route-route-point.controller';
+export * from './edge-point.controller';
