@@ -1,29 +1,29 @@
 import {
   Count,
   CountSchema,
-  Filter,
   FilterExcludingWhere,
   repository,
   Where,
 } from '@loopback/repository';
 import {
-  post,
-  param,
+  del,
   get,
   getModelSchemaRef,
+  param,
   patch,
+  post,
   put,
-  del,
   requestBody,
   response,
 } from '@loopback/rest';
-import {Point} from '../models';
+import Graph from 'node-dijkstra';
+import {BestRouteReq, Point} from '../models';
 import {PointRepository} from '../repositories';
 
 export class PointController {
   constructor(
     @repository(PointRepository)
-    public pointRepository : PointRepository,
+    public pointRepository: PointRepository,
   ) {}
 
   @post('/points')
@@ -47,14 +47,50 @@ export class PointController {
     return this.pointRepository.create(point);
   }
 
+  @post('/points/best-route')
+  @response(200, {
+    description: 'Find the best route',
+  })
+  async bestRoute(
+    @requestBody({
+      content: {
+        'application/json': {
+          schema: getModelSchemaRef(BestRouteReq, {
+            title: 'BestRouteReq',
+          }),
+        },
+      },
+    })
+    bestRouteReq: BestRouteReq,
+  ) {
+    const points = await this.pointRepository.find();
+    const route = new Graph();
+    points.forEach(point => {
+      const neighbors: {
+        [key: string]: number;
+      } = {};
+      point.edges.forEach(edge => {
+        neighbors[edge.pointId] = edge.weight;
+      });
+      route.addNode(point._id!, neighbors);
+    });
+    const bestRoute = route.path(bestRouteReq.origin, bestRouteReq.destination);
+    return this.pointRepository.find({
+      where: {
+        _id: {inq: bestRoute as string[]},
+      },
+      fields: {
+        edges: false,
+      },
+    });
+  }
+
   @get('/points/count')
   @response(200, {
     description: 'Point model count',
     content: {'application/json': {schema: CountSchema}},
   })
-  async count(
-    @param.where(Point) where?: Where<Point>,
-  ): Promise<Count> {
+  async count(@param.where(Point) where?: Where<Point>): Promise<Count> {
     return this.pointRepository.count(where);
   }
 
@@ -70,10 +106,10 @@ export class PointController {
       },
     },
   })
-  async find(
-    @param.filter(Point) filter?: Filter<Point>,
-  ): Promise<Point[]> {
-    return this.pointRepository.find(filter);
+  async find(): Promise<Point[]> {
+    return this.pointRepository.find({
+      fields: {edges: false},
+    });
   }
 
   @patch('/points')
@@ -106,7 +142,8 @@ export class PointController {
   })
   async findById(
     @param.path.string('id') id: string,
-    @param.filter(Point, {exclude: 'where'}) filter?: FilterExcludingWhere<Point>
+    @param.filter(Point, {exclude: 'where'})
+    filter?: FilterExcludingWhere<Point>,
   ): Promise<Point> {
     return this.pointRepository.findById(id, filter);
   }
