@@ -4,22 +4,19 @@ export * from './credentials.model';
 export * from './driver-points.model';
 export * from './driver-ubication.model';
 export * from './generate-otp.model';
+export * from './nearest-driver-request.model';
+export * from './notification.model';
 export * from './password-recovery.model';
 export * from './password-reset.model';
 export * from './point.model';
 export * from './pqrs.model';
 export * from './role.model';
 export * from './sign-up-credentials.model';
+export * from './trip-comment.model';
+export * from './trip-rating.model';
+export * from './trip.model';
 export * from './user.model';
 export * from './validate-otp.model';
 export * from './vehicle.model';
-export * from './route.model';
-export * from './trip.model';
-export * from './trip-comment.model';
-export * from './code-2-fa.model';
-export * from './notification.model';
-export * from './trip-rating.model';
-export * from './route-point.model';
-export * from './create-route-request.model';
-export * from './nearest-driver-request.model';
-export * from './points-for-dijkstra.model';
+export * from './edge.model';
+export * from './best-route-req.model';
