@@ -1,13 +1,12 @@
+export * from './code-2-fa.repository';
+export * from './driver-ubication.repository';
+export * from './notification.repository';
+export * from './point.repository';
 export * from './pqrs.repository';
 export * from './role.repository';
+export * from './trip-comment.repository';
+export * from './trip-rating.repository';
+export * from './trip.repository';
 export * from './user.repository';
 export * from './vehicle.repository';
-export * from './code-2-fa.repository';
-export * from './notification.repository';
-export * from './trip-rating.repository';
-export * from './driver-ubication.repository';
-export * from './point.repository';
-export * from './route.repository';
-export * from './trip-comment.repository';
-export * from './trip.repository';
-export * from './route-point.repository';
+export * from './edge.repository';
