@@ -1,6 +1,5 @@
-export * from './sendgrid.service';
 export * from './auth.service';
-export * from './users.service';
-export * from './twilio.service'
 export * from './cloudinary.service';
-export * from './route.service';
+export * from './sendgrid.service';
+export * from './twilio.service';
+export * from './users.service';
